requirements.txt
requests
sqlalchemy
python-dotenv
mysql-connector-python

